Merge commit 'b571e9391ca7959e1d0dbbad8ab93a6ae01ede8e' into dev

+++ runtime/observatory/lib/utils.dart
@@ -97,7 +97,7 @@ class Utils {
   }
 
   static String formatSize(bytesDynamic) {
-    int bytes = bytesDynamic;
+    int bytes = bytesDynamic.toInt();
     const int digits = 1;
     const int bytesPerKB = 1024;
     const int bytesPerMB = 1024 * bytesPerKB;
